added modulefactory, debugging instantiation with parameter
